kcqrs: add rpc support
Added RPC support to the library by introducing an HTTP layer over the
existing EventStore implementation.

The library provides a client which is using Google HTTP Client to
communicate with remote endpoint. The backend is not going to be added
to the library, so you are able to make it for your-self using the
library modules such as EventStore.

Fixes #4