kcqrs/eventstore: read all events from the event store
Added API for retrieving of all events from the event store. The GAE implementation of the EventStore uses an additional Entity to keep a index of all added events and is used for querying of the datastore.

The kcqrs-client is updated regarding this change and calls remote HTTP endpoint for retrieving of the events.