kcqrs: add support for enc/dec events into protobuf messages
Added API for encoding/decoding Events to and from protobuf messages.
This change allows client code to be able to use and protobuf messages
instead of only json ones.

The protobuf messages are having better versioning capatibilities
(property renames, property changes, adding of new props or deprecation
of existing) which is and a better option for new projects.